refactor(ffi-uniffi): split errors.rs into directory module (closes #29)
Mirror the bridge crate's error/ directory pattern (PR #53) for the
uniffi-side error layer. The flat errors.rs had grown to 532 lines —
~60% tests, ~40% source — making the file harder to scan than its
bridge-crate counterpart.

Split layout:
- errors/mod.rs   (31)  re-exports + module-level docstring
- errors/unlock.rs (145) UnlockError + From<FfiUnlockError> + tests
- errors/vault.rs  (379) VaultError + From<FfiVaultError> + tests

No semantic change: every type, variant, error message, From mapping
and test moved verbatim. lib.rs is unchanged — the existing
`pub use errors::{UnlockError, VaultError}` line already feeds uniffi
scaffolding's `crate::TypeName` paths from the new mod.

Resolves issue #29 in full. The remaining items from that issue's
original scope (PyO3 lib.rs split + bridge error.rs/vault.rs split)
were closed by intermediate PRs #43 / #53; this PR finishes the
uniffi-side split that those didn't touch.
